Output metadata & iota.go updates (#1494)
* - Separate output metadata into separate type in the JSON api

* - Added BelowMaxDepth to the ProtocolParameters

* - Removed TokenTag

* - Added protocol version to milestones

* - Update INX

* - Updating required inx-coordinator to 0.3

* - Updated tests and snapshot files

* - Update inx-indexer and inx-mqtt to 0.5 and inx-faucet to 0.4

* - Fixed value test

+++ pkg/model/participation/test/testenv.go
@@ -33,7 +33,7 @@ var (
 	seed4, _       = hex.DecodeString("bd6fe09d8a309ca309c5db7b63513240490109cd0ac6b123551e9da0d5c8916c4a5a4f817e4b4e9df89885ce1af0986da9f1e56b65153c2af1e87ab3b11dabb4")
 
 	MinPoWScore   = 100.0
-	BelowMaxDepth = 15
+	BelowMaxDepth = uint16(15)
 )
 
 type ParticipationTestEnv struct {
